Add child elements to sidebar grid item
Three child items added to sidebar: a logo div and two divs for a main nav
and a sub nav area for settings, privacy, etc.

The list styles for both nav areas are set to none. Margin was added to
each child item to give it spacing. Margin was added to list items in
each ul to give spacing between nav items.

A root variable of parent-margin now will set margins for other main
grid child items so that there is consistancy in the design going
forward.