Update HTML/CSS structure and styles for header section
The Header now is a grid display with 2 rows and 2 columns.
Within the header are four divs that contain grouped elements.
The first is the search icon and bar, then the notifications
for the user, followed by the user avatar and information,
and finally the buttons section.

They were each styled then with a display of flex and using padding
as well as margins to add spacing.

Buttons were styled to have oval borders.